+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This update also moves the Camera options from the Settings to the Camera page, 
 ### Features
 
  - Update InfluxDB database export/import to use new Enterprise-compatible backup format
+ - Add general camera options: stream height/width, hide last still, and hide last timelapse ([#703](https://github.com/kizniche/mycodo/issues/703))
  - Add picamera options: white balance, shutter speed, sharpness, iso, exposure mode, meter mode, and image effect ([#313](https://github.com/kizniche/mycodo/issues/313), [#703](https://github.com/kizniche/mycodo/issues/703))
 
 ### Miscellaneous